Move Tournament and Event entities into TournamentAggregate
