+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/EventGenderValidatorTests.cs
@@ -1,7 +1,8 @@
 ﻿using FluentAssertions;
+using TournamentManagement.Domain.TournamentAggregate;
 using Xunit;
 
-namespace TournamentManagement.Domain.UnitTests
+namespace TournamentManagement.Domain.UnitTests.TournamentAggregate
 {
 	public class EventGenderValidatorTests
 	{
