+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/RoundTests.cs
@@ -1,5 +1,6 @@
 ﻿using FluentAssertions;
 using System;
+using TournamentManagement.Domain.TournamentAggregate;
 using Xunit;
 
 namespace TournamentManagement.Domain.UnitTests
