+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/SetWinnerCalculatorTests.cs
@@ -1,5 +1,6 @@
 ﻿using FluentAssertions;
 using System;
+using TournamentManagement.Domain.TournamentAggregate;
 using Xunit;
