+++ src/TournamentManagement/TournamentManagement.Domain/TournamentAggregate/SetType.cs
@@ -1,4 +1,4 @@
-﻿namespace TournamentManagement.Domain
+﻿namespace TournamentManagement.Domain.TournamentAggregate
 {
 	public enum SetType
 	{
