+++ src/TournamentManagement/TournamentManagement.Domain/TournamentAggregate/Tournament.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 
-namespace TournamentManagement.Domain
+namespace TournamentManagement.Domain.TournamentAggregate
 {
 	public class Tournament : Entity<TournamentId>
 	{
